convert flats to fits files
The file chi_convert_flats.pro was created to convert existing master
flat files from the archaic wdsk/rdsk format to fits format. This was
implemented in CHIRON back in September 2013, but was only run on
existing data back to January 2013. chi_create_many_logs was modified
to run chi_convert_flats on nights from 120302 to 121231.